+++ src/plugins/score-plugin-avnd/Crousti/CpuFilterNode.hpp
@@ -327,8 +327,8 @@ struct GfxNode<Node_T> final
   oscr::ProcessModel<Node_T>& processModel;
   GfxNode(
       oscr::ProcessModel<Node_T>& element,
-      std::weak_ptr<Execution::ExecutionCommandQueue> q, Gfx::exec_controls ctls, int id,
-      const score::DocumentContext& ctx)
+      std::weak_ptr<Execution::ExecutionCommandQueue> q, Gfx::exec_controls ctls,
+      int64_t id, const score::DocumentContext& ctx)
       : CustomGfxNodeBase{ctx}
       , GpuControlOuts{std::move(q), std::move(ctls)}
       , processModel{element}
